Improvements from code review comments
