+++ csharp/src/Drivers/Apache/Hive2/HiveServer2Statement.cs
@@ -15,15 +15,19 @@
 * limitations under the License.
 */
 
+using System;
 using System.Threading.Tasks;
+using Apache.Arrow.Ipc;
 using Apache.Hive.Service.Rpc.Thrift;
 
 namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
 {
     public abstract class HiveServer2Statement : AdbcStatement
     {
-        protected HiveServer2Connection connection;
-        protected TOperationHandle? operationHandle;
+        private const int PollTimeMillisecondsDefault = 500;
+        private const int BatchSizeDefault = 50000;
+        protected internal HiveServer2Connection connection;
+        protected internal TOperationHandle? operationHandle;
 
         protected HiveServer2Statement(HiveServer2Connection connection)
         {
@@ -32,6 +36,59 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
 
         protected virtual void SetStatementProperties(TExecuteStatementReq statement)
         {
+        }
+
+        protected abstract IArrowArrayStream NewReader<T>(T statement, Schema schema) where T : HiveServer2Statement;
+
+        public override QueryResult ExecuteQuery() => ExecuteQueryAsync().AsTask().Result;
+
+        public override UpdateResult ExecuteUpdate() => ExecuteUpdateAsync().Result;
+
+        public override async ValueTask<QueryResult> ExecuteQueryAsync()
+        {
+            await ExecuteStatementAsync();
+            await PollForResponseAsync();
+            Schema schema = await GetSchemaAsync();
+
+            // TODO: Ensure this is set dynamically based on server capabilities
+            return new QueryResult(-1, NewReader(this, schema));
+        }
+
+        public override async Task<UpdateResult> ExecuteUpdateAsync()
+        {
+            const string NumberOfAffectedRowsColumnName = "num_affected_rows";
+
+            QueryResult queryResult = await ExecuteQueryAsync();
+            if (queryResult.Stream == null)
+            {
+                throw new AdbcException("no data found");
+            }
+
+            using IArrowArrayStream stream = queryResult.Stream;
+
+            // Check if the affected rows columns are returned in the result.
+            Field affectedRowsField = stream.Schema.GetFieldByName(NumberOfAffectedRowsColumnName);
+            if (affectedRowsField != null && affectedRowsField.DataType.TypeId != Types.ArrowTypeId.Int64)
+            {
+                throw new AdbcException($"Unexpected data type for column: '{NumberOfAffectedRowsColumnName}'", new ArgumentException(NumberOfAffectedRowsColumnName));
+            }
+
+            // If no altered rows, i.e. DDC statements, then -1 is the default.
+            long? affectedRows = null;
+            while (true)
+            {
+                using RecordBatch nextBatch = await stream.ReadNextRecordBatchAsync();
+                if (nextBatch == null) { break; }
+                Int64Array numOfModifiedArray = (Int64Array)nextBatch.Column(NumberOfAffectedRowsColumnName);
+                // Note: should only have one item, but iterate for completeness
+                for (int i = 0; i < numOfModifiedArray.Length; i++)
+                {
+                    // Note: handle the case where the affected rows are zero (0).
+                    affectedRows = (affectedRows ?? 0) + numOfModifiedArray.GetValue(i).GetValueOrDefault(0);
+                }
+            }
+
+            return new UpdateResult(affectedRows ?? -1);
         }
 
         protected async Task ExecuteStatementAsync()
@@ -53,7 +110,7 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
             TGetOperationStatusResp? statusResponse = null;
             do
             {
-                if (statusResponse != null) { await Task.Delay(500); }
+                if (statusResponse != null) { await Task.Delay(PollTimeMilliseconds); }
                 TGetOperationStatusReq request = new TGetOperationStatusReq(this.operationHandle);
                 statusResponse = await this.connection.Client.GetOperationStatus(request);
             } while (statusResponse.OperationState == TOperationState.PENDING_STATE || statusResponse.OperationState == TOperationState.RUNNING_STATE);
@@ -65,6 +122,10 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2
             TGetResultSetMetadataResp response = await this.connection.Client.GetResultSetMetadata(request);
             return SchemaParser.GetArrowSchema(response.Schema);
         }
+
+        protected internal int PollTimeMilliseconds { get; } = PollTimeMillisecondsDefault;
+
+        protected internal int BatchSize { get; } = BatchSizeDefault;
 
         public override void Dispose()
         {
